+++ src/A/tests/index.test.js
@@ -3,8 +3,9 @@
  */
 import React from 'react';
 import { shallow, mount } from 'enzyme';
-import BootstrapProvider from '../../BootstrapProvider';
+import BootstrapProvider from 'bootstrap-styled-provider/lib/BootstrapProvider';
 import A from '../index';
+
 
 const children = (<h1>Test</h1>);