+++ src/InputGroup/tests/InputGroupButton.test.js
@@ -3,8 +3,8 @@
  */
 import { shallow, mount } from 'enzyme';
 import React from 'react';
+import BootstrapProvider from 'bootstrap-styled-provider/lib/BootstrapProvider';
 import InputGroupButton from '../InputGroupButton';
-import BootstrapProvider from '../../BootstrapProvider';
 import Button from '../../Button';
 const children = (<h1>Test</h1>);