+++ src/Jumbotron/tests/index.test.js
@@ -4,7 +4,7 @@
 
 import { shallow, mount } from 'enzyme';
 import React from 'react';
-import BootstrapProvider from '../../BootstrapProvider';
+import BootstrapProvider from 'bootstrap-styled-provider/lib/BootstrapProvider';
 import Jumbotron from '../index';
 
 const children = <h1> test </h1>;