+++ src/Modal/tests/ModalBody.test.js
@@ -3,7 +3,7 @@
  */
 import { shallow, mount } from 'enzyme';
 import React from 'react';
-import BootstrapProvider from '../../BootstrapProvider';
+import BootstrapProvider from 'bootstrap-styled-provider/lib/BootstrapProvider';
 import ModalBody from '../ModalBody';
 
 const children = <p>Test</p>;