+++ src/Nav/tests/NavLink.test.js
@@ -4,7 +4,7 @@
 
 import { mount } from 'enzyme';
 import React from 'react';
-import BootstrapProvider from '../../BootstrapProvider';
+import BootstrapProvider from 'bootstrap-styled-provider/lib/BootstrapProvider';
 
 import NavLink from '../NavLink';